updating data and handling new data vlues in code

+++ Unity/Assets/client/scripts/Data/Specifications/Weapons/Assailments/Assailer.cs
@@ -6,12 +6,12 @@ namespace JunkyardDogs.Specifications
     public class Assailer : PhysicalSpecification
     {
         [SerializeField]
-        private int _damage;
+        private float _damage;
 
         [SerializeField]
         private Effect[] _effects;
 
-        public int Damage { get { return _damage; }
+        public float Damage { get { return _damage; }
             set { _damage = value; }
         }
         public Effect[] Effects { get { return _effects; } }
